+++ lifecycle/src/commonJvmMain/kotlin/com/hoc081098/solivagant/lifecycle/internal/AtomicReference.commonJvm.kt
@@ -1,7 +1,21 @@
-package com.hoc081098.solivagant.lifecycle.internal
+/*
+ * Copyright 2024 Petrus Nguyễn Thái Học
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 
-import java.util.concurrent.atomic.AtomicReference
+package com.hoc081098.solivagant.lifecycle.internal
 
 // TODO: https://youtrack.jetbrains.com/issue/KT-37316
 @Suppress("ACTUAL_WITHOUT_EXPECT") // internal expect is not matched with internal typealias to public type
-internal actual typealias AtomicReference<T> = AtomicReference<T>
+internal actual typealias AtomicReference<T> = java.util.concurrent.atomic.AtomicReference<T>