+++ navigation/src/commonMain/kotlin/com/hoc081098/solivagant/navigation/internal/StackEntry.kt
@@ -1,3 +1,35 @@
+/*
+ * Copyright 2021 Freeletics GmbH.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/*
+ * Copyright 2024 Petrus Nguyễn Thái Học
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package com.hoc081098.solivagant.navigation.internal
 
 import androidx.compose.runtime.Immutable
@@ -7,12 +39,11 @@ import com.hoc081098.solivagant.navigation.ContentDestination
 import com.hoc081098.solivagant.navigation.NavRoot
 import com.hoc081098.solivagant.navigation.NavRoute
 import dev.drewhamilton.poko.Poko
-import kotlin.jvm.JvmInline
 
 @Poko
 @Immutable
 internal class StackEntry<T : BaseRoute> private constructor(
-  val id: Id,
+  val id: StackEntryId,
   val route: T,
   val destination: ContentDestination<T>,
   val lifecycleOwner: StackEntryLifecycleOwner,
@@ -27,9 +58,6 @@ internal class StackEntry<T : BaseRoute> private constructor(
       is NavRoute -> true
       is NavRoot -> false
     }
-
-  @JvmInline
-  value class Id(val value: String)
 
   companion object {
     inline fun <T : BaseRoute> create(
@@ -46,7 +74,7 @@ internal class StackEntry<T : BaseRoute> private constructor(
         )
 
       return StackEntry(
-        id = Id(idGenerator()),
+        id = StackEntryId(idGenerator()),
         route = route,
         destination = destination,
         lifecycleOwner = StackEntryLifecycleOwner(